PredictiveProject/requirements.txt
certifi==2018.11.29
dj-database-url==0.5.0
Django==1.11.18
django-heroku==0.3.1
gunicorn==19.9.0
pipenv==2018.11.26
psycopg2==2.7.7
pytz==2018.9
virtualenv==16.4.1
virtualenv-clone==0.5.1
whitenoise==4.1.2
idna==2.8
numpy==1.16.2
pandas==0.24.1
pip==18.1
python-dateutil==2.8.0
requests==2.21.0
setuptools==40.8.0
six==1.12.0
urllib3==1.24.1
chardet==3.0.4
heroku==0.1.4
patsy==0.5.1
scikit-learn==0.20.3
scipy==1.2.1
sklearn==0.0
stats==0.1.2a0
statsmodels==0.9.0
django-filter==2.1.0
django-widget-tweaks==1.4.3
django-bootstrap3==11.0.0
